Introduce ProxyTextFile
Steps that write to output files cannot just write to normal file-like
objects because we need to switch to BytesIO-backed files when running in
parallel. The ProxyTextFile acts like a normal file, but is backed by a
BytesIO object.

The plan is to make it possible to create a Pipeline just once and not
have to recreate in each worker.