Convert the rest of the steps to use proxy files
We now no longer need to re-create the pipeline steps for each chunk.

This allows us to get rid of all the weird _add_untrimmed_filter() methods
on Pipeline etc.

We also get rid of "setting the filtering parameters by setting an attribute
on the pipeline" (pipeline.discard_trimmed = True).

The code where the demultiplexers create their output files is now within
the Demultiplexer classes, where it should be.

This is now much closer to what could become a usable API.